requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.40.0
django-countries==5.5
django-crispy-forms==1.7.2
django-environ==0.4.5
idna==2.8
oauthlib==3.1.0
Pillow==6.2.2
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
stripe==2.37.2
urllib3==1.25.6
gunicorn==20.0.4
whitenoise==5.2.0
opentelemetry-sdk==1.10.0
opentelemetry-api==1.10.0
opentelemetry-exporter-otlp-proto-http==1.10.0
opentelemetry-proto==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-django==0.29b0
opentelemetry-instrumentation-requests==0.29b0
opentelemetry-instrumentation-wsgi==0.29b0
